add a complete test for multi-phenomena volcano calculation

+++ openquakeplatform_ipt/test/examples_test.py
@@ -19,7 +19,7 @@ from openquakeplatform.settings import FILE_PATH_FIELD_DIRECTORY
 from openquake.moon import platform_get, TimeoutError
 
 PLA_ADMIN_ID = os.environ.get('GEM_PLA_ADMIN_ID', '1')
-
+_CLEAN_ALL = True
 #
 # TO RUN A SINGLE TEST:
 #
@@ -118,7 +118,10 @@ imt_examples = {
              'sfx': 'zip'},
             {'exa_id': 99, 'subtab_id': 3,
              'zipfile': 'EventBasedHazard.zip',
-             'sfx': 'zip'}
+             'sfx': 'zip'},
+            {'exa_id': 99, 'subtab_id': 4,
+             'zipfile': 'Volcano.zip',
+             'sfx': 'zip'},
         ]
     }
 }
@@ -151,7 +154,7 @@ def replicatetree(fm, to):
         if os.path.isdir(fm_item):
             if os.path.exists(to_item):
                 if os.path.isdir(to_item):
-                    continue
+                    replicatetree(fm_item, to_item)
                 else:
                     raise OSError("'%s' is not a directory" % to_item)
             else:
@@ -165,7 +168,7 @@ def replicatetree(fm, to):
 
 
 def zip_diff(filename1, filename2):
-    differs = False
+    differs = True
 
     z1 = zipfile.ZipFile(open(filename1, "rb"))
     z2 = zipfile.ZipFile(open(filename2, "rb"))
@@ -177,17 +180,31 @@ def zip_diff(filename1, filename2):
         return 1
     for zipentry in z1.infolist():
         if zipentry.filename not in z2.namelist():
-            diff = difflib.ndiff(z1.open(zipentry.filename),
-                                 z2.open(zipentry.filename))
-            delta = ''.join(x[2:] for x in diff
-                            if x.startswith('- ') or x.startswith('+ '))
-            if delta:
-                print("zip_diff: differ here: %s" % zipentry.filename)
-                differs = True
+            break
+        cont1 = z1.open(zipentry.filename).read()
+        cont2 = z2.open(zipentry.filename).read()
+
+        _, ext = os.path.splitext(zipentry.filename)
+        if ext.upper() == '.ZIP' or ext.upper() == '.HDF5':
+            if cont1 != cont2:
+                print("\nzip_diff: the files %s are binarily different\n" % (
+                    zipentry.filename,))
                 break
-    if not differs:
-        return 0
-    return 1
+        else:
+            delta = ''
+            diff = difflib.ndiff(list(z1.open(zipentry.filename)),
+                                 list(z2.open(zipentry.filename)))
+            delta = ''.join(x for x in diff
+                            if x.startswith('- ') or x.startswith('+ '))
+
+            if delta:
+                print("\nzip_diff: the files %s are different:\n%s\n" % (
+                    zipentry.filename, delta))
+                break
+    else:
+        differs = False
+
+    return (1 if differs else 0)
 
 
 def setup_module(module):
@@ -213,7 +230,7 @@ def setup_module(module):
 
 
 def teardown_module(module):
-    if _fpath_field_directory_old is not None:
+    if (_CLEAN_ALL and _fpath_field_directory_old is not None):
         shutil.rmtree(_fpath_field_directory)
         os.rename(_fpath_field_directory_old,
                   _fpath_field_directory)
@@ -350,14 +367,15 @@ class IptUploadTest(unittest.TestCase):
             " and @name='configuration_file']"
             "//div[starts-with(@id, 'cf_subtabs-') and @name='scenario']")
 
-        clean_all = pla.xpath_finduniq(
-            common + "//button[@type='submit' and @name='clean_all']")
-        clean_all.click()
+        if _CLEAN_ALL:
+            clean_all = pla.xpath_finduniq(
+                common + "//button[@type='submit' and @name='clean_all']")
+            clean_all.click()
 
-        confirm = pla.xpath_finduniq(
-            "//div[@class='ui-dialog-buttonset']//button["
-            "@type='button' and normalize-space(text())='Yes']")
-        confirm.click()
+            confirm = pla.xpath_finduniq(
+                "//div[@class='ui-dialog-buttonset']//button["
+                "@type='button' and normalize-space(text())='Yes']")
+            confirm.click()
 
         # show div with upload file
         up_file = os.path.join(os.path.dirname(__file__), 'data',
@@ -408,14 +426,15 @@ class IptExamplesTest(unittest.TestCase):
             " @name='configuration_file']"
             "//div[starts-with(@id, 'cf_subtabs-') and @name='scenario']")
 
-        clean_all = pla.xpath_finduniq(
-            common + "//button[@type='submit' and @name='clean_all']")
-        clean_all.click()
+        if _CLEAN_ALL:
+            clean_all = pla.xpath_finduniq(
+                common + "//button[@type='submit' and @name='clean_all']")
+            clean_all.click()
 
-        confirm = pla.xpath_finduniq(
-            "//div[@class='ui-dialog-buttonset']//button["
-            "@type='button' and normalize-space(text())='Yes']")
-        confirm.click()
+            confirm = pla.xpath_finduniq(
+                "//div[@class='ui-dialog-buttonset']//button["
+                "@type='button' and normalize-space(text())='Yes']")
+            confirm.click()
 
         #
         # populate uploaded file with template
